Make enum field of the Server Variable Object optional
According to the [OpenAPI
spec](https://spec.openapis.org/oas/v3.1.1.html#fixed-fields-4) the
`enum` field of the Server Variable object is optional (and when it's
present it must be a non-empty array).

Before this change, when parsing an OpenAPI document where a Server
Variable object is present without the `enum` field, this error would be
thrown:

```
type 'Null' is not a subtype of type 'Iterable<dynamic>'
package:open_api_forked/src/v3/server.dart 73:48
APIServerVariable.decode
```

This change makes it possible to parse such document and the resulting
`availableValues` is `null`.